refactor:fix project structure 2
